+++ packages/web/docs/src/content/graphql-api/_meta.ts
@@ -5,4 +5,5 @@ export default {
   'project-management': 'Manage Projects',
   'target-management': 'Manage Targets',
   'contract-management': 'Manage Contracts',
+  'unused-deprecated-schema': 'Lookup unused/deprecated Schema',
 };